+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][keep-a-changelog]
 
 ## [Unreleased]
 - Fix for deferred authentication failing with signature failure (frejaeid)
+- Implementing getAll (frejaorgid)
+- Fixed broken creation data return if created by other than SSN (frejaorgid)
 
 ## [0.1.8] (2020-11-24)